[Update] Complete fix of linting errors
