+++ src/dom-interface.js
@@ -1,26 +1,8 @@
 import makeRequest from './handler';
 
-const countries = require('i18n-iso-countries');
-countries.registerLocale(require('i18n-iso-countries/langs/en.json'));
 
 const selectQuery = query => document.querySelector(query);
 const searchIcon = selectQuery('.icon-wrap');
-
-const iconSwitch = weather => {
-  const icon = selectQuery('.wi');
-  const weatherDesc = selectQuery('.weather-desc');
-  icon.classList = `wi wi-owm-${weather.id}`;
-  weatherDesc.textContent = weather.description;
-};
-
-const checkWeather = data => {
-  const weatherIcon = selectQuery('.umbrella');
-  const advise = selectQuery('.advise>p');
-  if (!/cloud/.test(data)) {
-    weatherIcon.src = '../src/assets/fun.png';
-    advise.textContent = 'Go have fun.';
-  }
-};
 
 const getInput = () => {
   const searchIcon = selectQuery('.icon-wrap');
@@ -63,97 +45,5 @@ const load = () => {
   `;
 };
 
-const showCountry = (query, city) => {
-  const location = selectQuery('.location');
-  let quest;
-  if (query === 'US') quest = 'USA';
-  location.textContent = `${city || 'Unknown'}, ${quest
-    || countries.getName(query, 'en')
-    || ' Location'}`;
-};
 
-const switchLoader = () => {
-  const top = selectQuery('.top-ad');
-  const temp = selectQuery('.temp');
-  const weather = selectQuery('.weather-more');
-  const advise = selectQuery('.advise');
-  const load = selectQuery('.load-wrap');
-  top.style = 'opacity: 0';
-  temp.style = 'opacity: 0';
-  weather.style = 'opacity: 0';
-  advise.style = 'opacity: 0';
-  load.style = 'opacity: 1';
-
-  setTimeout(() => {
-    top.style = 'opacity: 1';
-    temp.style = 'opacity: 1';
-    weather.style = 'opacity: 1';
-    advise.style = 'opacity: 1';
-    load.style = 'opacity: 0';
-  }, 3500);
-};
-
-const renderData = (data, message) => {
-  const tempElem = selectQuery('.temp-h1');
-  if (!data) {
-    tempElem.style.fontSize = '1em';
-    tempElem.textContent = message;
-  }
-  switchLoader();
-
-  const mainData = data.main;
-  const { feels_like: feelsLike } = mainData;
-  const { temp } = mainData;
-
-  let newFeels = feelsLike * 1.8 + 32;
-  newFeels = Math.round((newFeels + Number.EPSILON) * 100) / 100;
-
-  let newTemp = temp * 1.8 + 32;
-  newTemp = Math.round((newTemp + Number.EPSILON) * 100) / 100;
-  const { country } = data.sys;
-  const city = data.name;
-  const { humidity } = mainData;
-  const wind = data.wind.speed;
-  const switchTemp = document.querySelector('input[type="checkbox"]');
-
-  const feelsElem = selectQuery('.feels-score');
-  const toCelsius = (tempElem, temp, Feels) => {
-    tempElem.textContent = temp;
-    document.body.style.setProperty('--farendisplay', 'none');
-    document.body.style.setProperty('--celdisplay', 'initial');
-    feelsElem.innerHTML = `${Feels}<span class="togg">C</span>`;
-  };
-
-  const toFaren = (tempElem, newTemp, like) => {
-    tempElem.textContent = newTemp;
-    document.body.style.setProperty('--celdisplay', 'none');
-    document.body.style.setProperty('--farendisplay', 'initial');
-    feelsElem.innerHTML = `${like}<span class="togg">F</span>`;
-  };
-
-  if (switchTemp.checked) {
-    toFaren(tempElem, newTemp, newFeels);
-  } else {
-    toCelsius(tempElem, temp, feelsLike);
-  }
-
-  showCountry(country, city);
-  tempElem.style.fontSize = '8em';
-
-  iconSwitch(data.weather[0]);
-  selectQuery('.humid-score').textContent = `${humidity}%`;
-  selectQuery('.wind-score').textContent = `${wind}mph`;
-
-  switchTemp.addEventListener('change', e => {
-    if (e.target.checked) {
-      toFaren(tempElem, newTemp, newFeels);
-    } else {
-      toCelsius(tempElem, temp, feelsLike);
-    }
-  });
-  checkWeather(data.weather[0].description);
-};
-
-export {
-  animeSearch, load, showCountry, renderData, getInput,
-};
+export { animeSearch, load, getInput };
