+++ src/handler.js
@@ -1,4 +1,4 @@
-import { renderData } from './dom-interface';
+import renderData from './load-data';
 
 let newMap;
 let mapObj;
@@ -6,8 +6,38 @@ let mapObj;
 const loadGoogleMapApi = require('load-google-maps-api');
 
 const selectQuery = query => document.querySelector(query);
+const makeRequest = async (query, check) => {
+  const key = process.env.OPEN_WEATHER_API_KEY;
+  let strQuery;
 
-const googleMaps = (lat, long) => {
+  switch (check) {
+    case 'location':
+      query = query.split(' ');
+      strQuery = `http://api.openweathermap.org/data/2.5/weather?lat=${query[0]}&lon=${query[1]}&units=metric&APPID=${key}`;
+      break;
+    default:
+      strQuery = `http://api.openweathermap.org/data/2.5/weather?q=${query}&units=metric&APPID=${key}`;
+      break;
+  }
+  if (!strQuery) return;
+
+  try {
+    const response = await fetch(strQuery, {
+      mode: 'cors',
+    });
+    const data = await response.json();
+    renderData(data);
+    new mapObj.Marker({
+      position: { lat: data.coord.lat, lng: data.coord.lon },
+      map: newMap,
+    });
+    newMap.panTo({ lat: data.coord.lat, lng: data.coord.lon });
+  } catch (error) {
+    renderData(false, error.message);
+  }
+};
+
+const googleMaps = (lat, long, request) => {
   const mapElem = selectQuery('#map');
   loadGoogleMapApi({ key: process.env.GOOGLE_MAP_API_KEY }).then(map => {
     mapObj = map;
@@ -29,42 +59,7 @@ const googleMaps = (lat, long) => {
       mapCreated.panTo(e.latLng);
     });
   });
+  if (request === true) makeRequest(`${lat} ${long}`, 'location');
 };
 
-
-const makeRequest = async (query, check, mapPin) => {
-  const key = process.env.OPEN_WEATHER_API_KEY;
-  let strQuery;
-
-  switch (check) {
-    case 'location':
-      query = query.split(' ');
-      strQuery = `http://api.openweathermap.org/data/2.5/weather?lat=${query[0]}&lon=${query[1]}&units=metric&APPID=${key}`;
-      break;
-    default:
-      strQuery = `http://api.openweathermap.org/data/2.5/weather?q=${query}&units=metric&APPID=${key}`;
-      break;
-  }
-  if (!strQuery) return;
-
-  try {
-    const response = await fetch(strQuery, {
-      mode: 'cors',
-    });
-    const data = await response.json();
-    renderData(data);
-    if (mapPin === true) {
-      googleMaps(data.coord.lat, data.coord.lon);
-    } else {
-      new mapObj.Marker({
-        position: { lat: data.coord.lat, lng: data.coord.lon },
-        map: newMap,
-      });
-      newMap.panTo({ lat: data.coord.lat, lng: data.coord.lon });
-    }
-  } catch (error) {
-    renderData(false, error.message);
-  }
-};
-
-export { makeRequest, googleMaps };
+export default googleMaps;
